+++ src/core/store/BIM/context.tsx
@@ -25,6 +25,7 @@ const initialState = {
     splatIds: [],
     floorplans: [],
     grid: null  ,
+    fragmentsStarted: false,
     buildingModel: {
       bimFile: null,
       building: null,